requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
nltk = 3.7.0
vaderSentiment = 3.3.2